Use private generic type for drop behavior
This patch avoids the thread-local to choose dropping behavior by
instead typecasting among private generic types for the inner value
type. It's not clear if that's actually sound:
https://github.com/rust-lang/unsafe-code-guidelines/issues/35#issuecomment-735491663

The current implementation does not compile due to a conflict between
the need for the concrete drop behavior types to be private and the need
for the (public) bound:

    Aliased<T, crate::aliasing::NoDrop>: Borrow<_>

I don't quite know how to work around that yet. What we really want is
to write:

    for<U: DropBehavior> Aliased<T, U>: Borrow<_>

But I don't think that's current expressible? Maybe through some more
trait magic?